Update to use kingpin for better cli help
